refactoring(searchBarAutocomplete): extract SearchBarInputComponent
